Move ContentLoader and PixelStreamWindowManager to control
