+++ tide/master/control/PixelStreamWindowManager.h
